Do not publish dependency management with POMs
Prior to this commit, the POMs and BOM produced by this build would also
publish dependency management information for artifacts that are not
provided by RSocket. This is problematic since this would
provide/enforce opinions about unrelated dependencies such as test and
logging dependencies.

This commit configures the dependency-management Gradle plugin to not
publish dependency management information with the generated POMs.

It also uses the java-platform Gradle plugin to produce a proper BOM
that does not contain dependencies, publishes dependency management
information for RSocket artifacts only, and does not provide any scope
opinion about those.

This commit also polishes the main build file and:

* uses the Netty BOM
* upgrades JUnit to its latest version and uses its BOM
* refactors the Gradle build files to avoid duplication

Signed-off-by: Brian Clozel <bclozel@pivotal.io>