Fix Gradle publication configuration
Prior to this change, the Gradle publication configuration would be
applied to the wrong level of projects (subprojects of subprojects!),
resulting in no artifacts being published to the repository.

This commit updates the JFrog artifactory Gradle plugin version and
fixes the application of the publication configuration to align with
what's been done with artifactory and bintray configurations.

Signed-off-by: Brian Clozel <bclozel@pivotal.io>