reduce package cycles (#774)
* Remove package cycle between resume and internal

Move ClientSetup and ServerSetup are closely associated with the
RSocketFactory implementations. Moving them into core where they are
used from breaks a package cycle between resume and internal.

ClientSetup is straight-forward and folds easily into
DefaultClientRSocketFactory.

ConnectionUtil is closely associated with ServerSetup and has been
folded into it.

Signed-off-by: Rossen Stoyanchev <rstoyanchev@pivotal.io>

* Remove package cycle between lease and exceptions

MissingLeaseException is moved to the lease package next to the
contracts it is declared in breaking a cycle between lease and
exceptions. This is an API breakage but I the impact should be
minor (none? the chances of something handling it explicitly) and
hence an acceptable trade-off.

Signed-off-by: Rossen Stoyanchev <rstoyanchev@pivotal.io>

* Add RSocketProtocolException in top-level package

Replaces (now deprecated) RSocketException from the exceptions package
which can lead to cycles (e.g. with frame package).

Other minor refinements:

- error code validation
- shared errorCode() implementation
- error code in toString()
- avoid NPE for null message, it's better to keep the original exception
  and null is allowed by getMessage().

Signed-off-by: Rossen Stoyanchev <rstoyanchev@pivotal.io>

* Deprecate ErrorType

This change deprecates ErrorType as the same constants are already
declared in ErrorFrameFlyweight and there is hierarchy of exceptions
with one exception per error code.

Signed-off-by: Rossen Stoyanchev <rstoyanchev@pivotal.io>