Deprecate AbstractRSocket and provide alternative
This change deprecates AbstractRSocket in favor of implementing RSocket
directly. In turn RSocket has default methods for the basic
interactions while onClose now returns Mono.never() by default and has
to be implemented anyway if it is to something more meaningful.

In addition SocketAcceptor provides static factor methods that provide
shortcuts for handling one type of interaction with a lambda syntax.

Signed-off-by: Rossen Stoyanchev <rstoyanchev@pivotal.io>