Remove use of unnecessary encodeLength/decodeLength flags
The flags were introduced in #585 such that when fragmentation is
enabled, it takes over the encoding and decoding of the frame length
instead of the underlying transport.

However this doesn't seem to have any obvious purpose. On the sending
side, fragmentation breaks down the frames and then the transport can
add the length just as well. On the receiving side, the transport is
first anyway.

It's better to simplify this and leave it to transports to decide this
whether a length is needed and expected.

Signed-off-by: Rossen Stoyanchev <rstoyanchev@pivotal.io>