Updates to WebSocketHeaders example
First, the sample no longer needs to apply the
ReassemblyDuplexConnection wrapper following d8da87.

Another problem is that it was incorrectly trying to send a WebSocket
status from an HTTP response code which was rejected by Reactor Netty.

Arguably unauthorized upgrades should be rejected during the initial
handshake, so now uses an HTTP request handler instead of a WebSocket
handler disallows the upgrade if not authorized.

Signed-off-by: Rossen Stoyanchev <rstoyanchev@pivotal.io>