update versions and fixes memleak in UnboundedProcessor (#1106)
This PR updates dependencies and makes minor modifications to UnboundedProcessor due to repeating failures of UnboundedProcessorJCStreassTest, which started reproducing some unspotted issues.

Motivation:

UnboundedProcessor is a critical component in the RSocket-Java ecosystem and must work properly. After analysis of its internal state machine, it was spotted that sometimes:

The request may not be delivered due to natural concurrency
The terminal signal may not be delivered since it checks for demand which might be consumed already (due to natural concurrency)
The final value could be delivered violating reactive-streams spec
Modifications:

This PR adds a minimal set of changes, preserving old implementation but eliminating the mentioned bugs

---------

Signed-off-by: Oleh Dokuka <oleh.dokuka@icloud.com>